# dataBaseServer/src/main/webapp/swaggerUI/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the API documentation for",
    "version" : "v1",
    "title" : "API"
  },
  "host" : "104.248.35.208:9090/dataBaseServer",
  "basePath" : "/api",
  "schemes" : [ "http" ],
  "paths" : {
    "/eyetrackerRaw" : {
      "post" : {
        "summary" : "Transfer eyetracking data",
        "description" : "Create a new data transfer for eyetracking data",
        "operationId" : "postEytracker",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Create a new data transfer for eyetracking data",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/EyeTrackerMessage"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Data transfer successful"
          },
          "400" : {
            "description" : "Error, bad request."
          }
        }
      }
    },
    "/shimmerRaw" : {
      "post" : {
        "summary" : "Transfer ShimmerMessage data",
        "description" : "Create a new data transfer for ShimmerMessage data",
        "operationId" : "postShimmer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Data transfer",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ShimmerMessage"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Data transfer is successfull. Message recieved"
          },
          "400" : {
            "description" : "Data transfer did not happen."
          }
        }
      }
    }
  },
  "definitions" : {
    "EyeTrackerMessage" : {
      "type" : "object",
      "properties" : {
        "device" : {
          "type" : "string"
        },
        "features" : {
          "type" : "string"
        },
        "data" : {
          "type" : "string"
        }
      }
    },
    "ShimmerMessage" : {
      "type" : "object",
      "properties" : {
        "device" : {
          "type" : "string"
        },
        "features" : {
          "type" : "string"
        },
        "data" : {
          "type" : "string"
        }
      }
    }
  }
}